Moved content types into internal package
